Add event_manager_get_permalink filter
Example use case: I have a single "my-account" page that contains all of the dashboards, accessible via tabs and subpaths like ".../my-account/organizer". I want to set the organizer_dashboard permalink to ".../my-account/organizer", but that's not possible, because it is not its own page.

Maybe it would be better to change how the options work, so you can set either a page id or a permalink in the settings, but that would be a lot of work. So this filter just gives developers something to work with.