Update wp-event-manager-writepanels.php
This fixes a bug that prevents you from deleting the contents of an additional field of type WP Editor.  As the code stands, you can create an additional field of type WP Editor and add or edit the detail in that field.  But if you later have no need for that extra field on an event, you can't clear the field because the current code does doesn't have any provision todo this.  This small update fixes that bug.